# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RUMIN API",
    "summary": "Financial intelligence and economic simulation platform.",
...
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ingestion-jobs": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "List ingestion jobs",
        "description": "Every retrieval or import, newest first, with what it did. A job's status is derived from its targets: one that only partly succeeded is `partially_failed`.",
        "operationId": "list_jobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "dataset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-z0-9][a-z0-9_-]{1,63}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Id"
            }
          },
          {
            "name": "provider_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-z0-9][a-z0-9-]{1,63}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPage"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingestion-jobs/{job_id}": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Get an ingestion job",
        "description": "The job with each target's outcome, issue counts by rule, and the responses or files it stored.",
        "operationId": "get_job",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/source-captures/{capture_id}": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Get a stored source capture",
        "description": "Metadata of a stored response or file: where it came from, when, its size and SHA-256. The stored bytes are not served by the API.",
        "operationId": "get_capture",
        "parameters": [
          {
            "name": "capture_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 4611686018427387904,
              "minimum": 1,
              "title": "Capture Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureRead"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data-quality/issues": {
      "get": {
        "tags": [
          "data quality"
        ],
        "summary": "List data-quality issues",
        "description": "Rejected records (with what the source sent), flagged values and notes, newest first. Issues are recorded per job.",
        "operationId": "list_issues",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          },
          {
            "name": "series_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-z0-9][a-z0-9-]{2,95}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Series Id"
            }
          },
          {
            "name": "instrument_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-z0-9][a-z0-9-]{2,95}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Instrument Id"
            }
          },
          {
            "name": "rule",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-z_]{2,64}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueSeverity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueOutcome"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Maximum number of items to return.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of items to return."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000,
              "minimum": 0,
              "description": "Items to skip.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Items to skip."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuePage"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data-quality/rules": {
      "get": {
        "tags": [
          "data quality"
        ],
        "summary": "List data-quality rules",
        "description": "Every rule applied during ingestion, what it checks and what happens to a record that fails it.",
        "operationId": "list_rules",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleRead"
                  },
                  "type": "array",
                  "title": "Response List Rules"
                }
              }
            }
          },
          "422": {
            "description": "The request contains invalid values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (details are logged).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaptureKind": {
        "type": "string",
        "enum": [
          "http_response",
          "file"
        ],
        "title": "CaptureKind"
      },
      "CaptureRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "kind": {
            "$ref": "#/components/schemas/CaptureKind"
          },
          "locator": {
            "type": "string",
            "title": "Locator",
            "description": "Sanitised URL, or the imported file's name."
          },
          "request_params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Params"
          },
          "http_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Status"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "body_stored": {
            "type": "boolean",
            "title": "Body Stored"
          },
          "provider_last_updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Last Updated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "provider_id",
          "kind",
          "locator",
          "request_params",
          "http_status",
          "received_at",
          "content_type",
          "size_bytes",
          "sha256",
          "body_stored",
          "provider_last_updated"
        ],
        "title": "CaptureRead",
        "description": "Metadata of bytes RUMIN received. The body itself is not served by the API."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code.",
            "examples": [
              "not_found"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable summary, safe to display."
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Correlates the error with server logs (X-Request-ID)."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Every non-2xx response from the API uses this envelope."
      },
...
      "IssueCount": {
        "properties": {
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "severity": {
            "$ref": "#/components/schemas/IssueSeverity"
          },
          "outcome": {
            "$ref": "#/components/schemas/IssueOutcome"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "severity",
          "outcome",
          "count"
        ],
        "title": "IssueCount"
      },
      "IssueOutcome": {
        "type": "string",
        "enum": [
          "rejected",
          "flagged",
          "noted"
        ],
        "title": "IssueOutcome",
        "description": "What happened to the record the issue is about."
      },
      "IssuePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IssueRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total number of items matching the query."
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "IssuePage",
        "description": "A page of data-quality issues, newest first."
      },
...
      "IssueSeverity": {
        "type": "string",
        "enum": [
          "error",
          "warning",
          "info"
        ],
        "title": "IssueSeverity"
      },
      "JobDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "trigger": {
            "$ref": "#/components/schemas/JobTrigger"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "What was requested. Never holds secrets."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "items_total": {
            "type": "integer",
            "title": "Items Total"
          },
          "items_succeeded": {
            "type": "integer",
            "title": "Items Succeeded"
          },
          "items_failed": {
            "type": "integer",
            "title": "Items Failed"
          },
          "items_skipped": {
            "type": "integer",
            "title": "Items Skipped"
          },
          "records_received": {
            "type": "integer",
            "title": "Records Received"
          },
          "records_new": {
            "type": "integer",
            "title": "Records New"
          },
          "records_revised": {
            "type": "integer",
            "title": "Records Revised"
          },
          "records_unchanged": {
            "type": "integer",
            "title": "Records Unchanged"
          },
          "records_missing": {
            "type": "integer",
            "title": "Records Missing",
            "description": "Accepted records without a value (a subset)."
          },
          "records_rejected": {
            "type": "integer",
            "title": "Records Rejected"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count"
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count"
          },
          "bytes_received": {
            "type": "integer",
            "title": "Bytes Received"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobItemRead"
            },
            "type": "array",
            "title": "Items"
          },
          "issue_counts": {
            "items": {
              "$ref": "#/components/schemas/IssueCount"
            },
            "type": "array",
            "title": "Issue Counts"
          },
          "captures": {
            "items": {
              "$ref": "#/components/schemas/CaptureRead"
            },
            "type": "array",
            "title": "Captures"
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider_id",
          "dataset_id",
          "trigger",
          "parameters",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "heartbeat_at",
          "items_total",
          "items_succeeded",
          "items_failed",
          "items_skipped",
          "records_received",
          "records_new",
          "records_revised",
          "records_unchanged",
          "records_missing",
          "records_rejected",
          "warning_count",
          "error_count",
          "request_count",
          "bytes_received",
          "error_summary",
          "items",
          "issue_counts",
          "captures"
        ],
        "title": "JobDetail"
      },
...
      "JobPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total number of items matching the query."
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "JobPage",
        "description": "A page of ingestion jobs, newest first."
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "trigger": {
            "$ref": "#/components/schemas/JobTrigger"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters",
            "description": "What was requested. Never holds secrets."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heartbeat At"
          },
          "items_total": {
            "type": "integer",
            "title": "Items Total"
          },
          "items_succeeded": {
            "type": "integer",
            "title": "Items Succeeded"
          },
          "items_failed": {
            "type": "integer",
            "title": "Items Failed"
          },
          "items_skipped": {
            "type": "integer",
            "title": "Items Skipped"
          },
          "records_received": {
            "type": "integer",
            "title": "Records Received"
          },
          "records_new": {
            "type": "integer",
            "title": "Records New"
          },
          "records_revised": {
            "type": "integer",
            "title": "Records Revised"
          },
          "records_unchanged": {
            "type": "integer",
            "title": "Records Unchanged"
          },
          "records_missing": {
            "type": "integer",
            "title": "Records Missing",
            "description": "Accepted records without a value (a subset)."
          },
          "records_rejected": {
            "type": "integer",
            "title": "Records Rejected"
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count"
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count"
          },
          "bytes_received": {
            "type": "integer",
            "title": "Bytes Received"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider_id",
          "dataset_id",
          "trigger",
          "parameters",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "heartbeat_at",
          "items_total",
          "items_succeeded",
          "items_failed",
          "items_skipped",
          "records_received",
          "records_new",
          "records_revised",
          "records_unchanged",
          "records_missing",
          "records_rejected",
          "warning_count",
          "error_count",
          "request_count",
          "bytes_received",
          "error_summary"
        ],
        "title": "JobRead"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "completed_with_warnings",
          "partially_failed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "description": "Lifecycle of an ingestion job. The final status always reflects what happened:\n\n* ``completed`` — every target succeeded; nothing was rejected or flagged.\n* ``completed_with_warnings`` — every target succeeded, but some records were\n  rejected or flagged, or a target returned no data.\n* ``partially_failed`` — some targets failed and others succeeded.\n* ``failed`` — no target succeeded, or the job could not run.\n* ``cancelled`` — stopped before finishing (e.g. interrupted from the keyboard)."
      },
...
      "JobTrigger": {
        "type": "string",
        "enum": [
          "cli"
        ],
        "title": "JobTrigger",
        "description": "Who started a job. Phase 2 starts jobs from the command line only."
      },
...
      "RuleRead": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "applies_to": {
            "type": "string",
            "enum": [
              "economic",
              "price",
              "both"
            ],
            "title": "Applies To"
          },
          "severity": {
            "$ref": "#/components/schemas/IssueSeverity"
          },
          "outcome": {
            "$ref": "#/components/schemas/IssueOutcome"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "code",
          "applies_to",
          "severity",
          "outcome",
          "description"
        ],
        "title": "RuleRead"
      },
...
    }
  },
...
}